+++ packages/richtext/src/index.ts
@@ -1,0 +1,1 @@
+export { getRichTextHtml } from './utils/parse-tiptap-json';